Move filter components to filter folder
