+++ src/Filters/atoms/FilterButton.tsx
@@ -1,12 +1,12 @@
-import React, { ButtonHTMLAttributes} from 'react';
+import React, { ButtonHTMLAttributes } from 'react';
 import styled, { css } from 'styled-components';
-import { TypeButtonSizes } from '..';
+import { TypeButtonSizes } from '../../Form';
 import { resetButtonStyles } from '../../common';
 import { dimensions } from '../../themes/common';
 import Icon, { IconWrapper } from '../../Icons/Icon';
 
 
-const StyledButton = styled.button<{size: TypeButtonSizes, fontSize: string,  isOpen?: boolean, hasFlipArrow?: boolean }>`
+const StyledButton = styled.button<{ size: TypeButtonSizes, fontSize: string, isOpen?: boolean, hasFlipArrow?: boolean }>`
   ${resetButtonStyles};
   background-color: hsl(200, 23%, 97%);
   border-radius: 3px;
@@ -17,11 +17,11 @@ const StyledButton = styled.button<{size: TypeButtonSizes, fontSize: string,  is
   ${({ theme, size, fontSize }) => theme && css`
     font-family: ${theme.fontFamily.ui};
     font-size: ${fontSize};
-    height: ${theme.dimensions.filterBar.button[ size ].height};
+    height: ${theme.dimensions.filterBar.button[size].height};
     transition: opacity ${theme.animation.speed.normal} ${theme.animation.easing.primary.easeOut};
 
     ${IconWrapper} {
-      padding: ${theme.dimensions.filterBar.button[ size ].iconPadding};
+      padding: ${theme.dimensions.filterBar.button[size].iconPadding};
       display: flex;
       align-items: center;
 
@@ -70,7 +70,7 @@ const InnerContainer = styled.div`
  * value of that might be replaced latter with a theme update.
  */
 const getFontSize = (size: TypeButtonSizes): string => {
-  switch(size) {
+  switch (size) {
     case 'xsmall':
       return '12px';
 
@@ -86,14 +86,14 @@ const getFontSize = (size: TypeButtonSizes): string => {
     default:
       return '12px';
   }
-}
+};
 
 const ArrowIcon = styled(Icon)``;
 
-const ButtonText = styled.div<{size: TypeButtonSizes, hasFlipArrow: boolean}>`
-  ${({theme, hasFlipArrow, size}) => hasFlipArrow
-  ? `padding-right: ${theme.dimensions.filterBar.button[ size ].textPaddingWithArrow};`
-  : `padding-right: ${theme.dimensions.filterBar.button[ size ].textPaddingNoArrow};`
+const ButtonText = styled.div<{ size: TypeButtonSizes, hasFlipArrow: boolean }>`
+  ${({ theme, hasFlipArrow, size }) => hasFlipArrow
+    ? `padding-right: ${theme.dimensions.filterBar.button[size].textPaddingWithArrow};`
+    : `padding-right: ${theme.dimensions.filterBar.button[size].textPaddingNoArrow};`
   };
 `;
 
@@ -120,13 +120,13 @@ const FilterButton: React.FC<IFilterButton> = ({
   const fontSize = getFontSize(size);
 
   return (
-    <StyledButton type='button' {...props} {...{size, fontSize, isOpen, hasFlipArrow}}>
+    <StyledButton type='button' {...props} {...{ size, fontSize, isOpen, hasFlipArrow }}>
       <InnerContainer>
         <Icon
           icon={icon}
           size={iconSize}
         />
-        <ButtonText {...{size, hasFlipArrow}} >{children}</ButtonText>
+        <ButtonText {...{ size, hasFlipArrow }}>{children}</ButtonText>
         {hasFlipArrow && <ArrowIcon icon={isOpen ? 'Up' : 'Down'} size={arrowSize} />}
       </InnerContainer>
     </StyledButton>
