+++ src/Filters/index.ts
@@ -1,7 +1,19 @@
 import DatePicker from './molecules/DatePicker';
 import FilterDropdownContainer from './atoms/FilterDropdownContainer';
+import FilterButton from './atoms/FilterButton';
+import FilterDropdown from './molecules/FilterDropdown';
+import FilterInputs, { ISearchFilter, IFilterInputs, IFilterDropdownExt } from './molecules/FilterInputs';
+import FiltersResults, { IFilterLabel } from './molecules/FiltersResults';
+import FilterBar, { IFilterDropdownConfig } from './organisms/FilterBar';
 
 export {
     DatePicker,
-    FilterDropdownContainer
+    FilterDropdownContainer,
+    FilterButton,
+    FilterDropdown,
+    FilterInputs,
+    FiltersResults,
+    FilterBar
 };
+
+export type { ISearchFilter, IFilterInputs, IFilterDropdownExt, IFilterLabel, IFilterDropdownConfig };
