+++ src/Misc/index.tsx
@@ -1,9 +1,7 @@
 import Tag, { ITag } from './atoms/Tag';
 import TagList, { ITagList, TagListWrapper } from './molecules/TagList';
 import BasicSearchInput from './atoms/BasicSearchInput';
-import FilterInputs, { ISearchFilter, IFilterDropdownExt } from './molecules/FilterInputs';
-import FiltersResults, { IFilterLabel } from './molecules/FiltersResults';
-import FilterBar, { IFilterDropdownConfig } from './organisms/FilterBar';
+
 import MediaBox from './atoms/MediaBox';
 
 export {
@@ -12,9 +10,6 @@ export {
   TagListWrapper,
   MediaBox,
   BasicSearchInput,
-  FilterInputs,
-  FiltersResults,
-  FilterBar,
 };
 
-export type { ITagList, ITag, ISearchFilter, IFilterDropdownExt, IFilterLabel, IFilterDropdownConfig };
+export type { ITagList, ITag};
